bug fix on type assertions

+++ pkg/handler/merchantHandler.go
@@ -58,7 +58,7 @@ func (handler *MerchantHandler) GetOmzetReport(c *fiber.Ctx) error {
 
 	userId, _ := payload.Get("user_id")
 	result := make([]entity.MerchantOmzet, 0)
-	if result, err = handler.merchantSrv.FindOmzetReportNovember(userId.(int64), pagination); err != nil {
+	if result, err = handler.merchantSrv.FindOmzetReportNovember(int64(userId.(float64)), pagination); err != nil {
 		response := fiber.Map{
 			"result": nil,
 			"error":  err.Error(),
